+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `--all` flag to `gantry list` to search all experiments, not just ones submitted through Gantry.
+
 ## [v2.0.2](https://github.com/allenai/beaker-gantry/releases/tag/v2.0.2) - 2025-05-05
 
 ### Fixed